% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.r
\name{calc.geweke}
\alias{calc.geweke}
\title{Calculate maximum absolute value of Geweke diagnostic across all chains (inner function)}
\usage{
calc.geweke(samples)
}
\arguments{
\item{samples}{coda::mcmc.list object}
}
\value{
A numeric value of Geweke diagnostic (as a rule, values less than 1.64 considered acceptable if the chains runs long enough)
}
\description{
Calculate maximum absolute value of Geweke diagnostic across all chains (inner function)
}
