% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funs.r
\name{calc.r.sq}
\alias{calc.r.sq}
\title{Calculate R-squared (inner function)}
\usage{
calc.r.sq(y, mu.rep)
}
\arguments{
\item{y}{A numeric vector of outcomes.}

\item{mu.rep}{A numberic vector of fitted values.}
}
\value{
An r-squared statistic calculated using \code{y} and \code{mu.rep}.
}
\description{
Calculate R-squared (inner function)
}
\examples{
calc.r.sq(c(3,2,4), c(1,2,0))
}
